This code predicts the results for the training dataset using the trained clf model:
clf.predict(X_train):
The clf model uses the features in X_train to predict the labels for each sample in the training set.

y_pred_train:
The predicted labels for the training set are stored in this variable. These predictions are typically used to assess the model's performance on the training data.